Allow Buffer NextWriter While Socket In Upgrading Process
This additional code allow buffering of writer when socket in
upgrading process and will write the buffer after the socket
upgrade complete. The whole socket connection write and read is
blocking each others, for example, if server receive an event
which if emit another event. The latter may fail if socket is in
upgrading no matter how much delay the emit wait for the
upgrading process because emit process itself blocked the reading
of socket upgraded packet. The problem can easily be seen when
two requests with different namespace connect to server when one
of connection is choosen to upgrade to websocket, the other
'connect' event blocked the former emit progress because the
'upgraded' packet can not be readed until 'connect' event packet
is been read.