Allow Buffer NextWriter While Socket In Upgrading Process
This additional code allow buffering of writer when socket in
upgrading process and will write the buffer after the socket
upgrade complete. The whole socket connection write and read is
blocking each others, for example, if server receive an event
which inside it emit another event (the emit must finish then
only next read can proceed). The latter may fail if socket is in
upgrading no matter how much delay the emit wait for the
upgrading process. This can happen when there is a message packet
send to server occurs before upgrading process start because this
message packet need to be synchronously read from socket loop
which is blocked by emitting process. Without the read of the
message packet (chan isn't read by socket loop), websocket can't
continue to read the upgraded packet and thus the socket always
in upgrading process. The problem can easily be seen when two
requests with different namespace connect to server when one of
connection is choosen to upgrade to websocket, the other 'connect'
event blocked the former emit progress (caused it return error
'upgrading') because the 'upgraded' packet can not be readed until
'connect' event packet is read which can only do so after emit
progress complete, thus a cyclic dead loop.

Updated existing tests so it can always pass go test with race
flag. Also added test for write to socket that is in upgrading
process and should not generate error.