Clear Pending NextWriter With Blocking When Buffer Not Close
There is one flaw in previous commit during clearing pending
writer buffer. If the pending writer buffer is not closed yet
then the clear pending writer buffer loop will keep on looping
until the buffer write is closed (accidental flaw as it was meant
to be coded to skip the buffer when it is not closed but always
getting Front from list.List object will loop and wait till buffer
is closed). That way may not be good for CPU cycle usage. This
commit run the clear pending writer buffer and will block when
buffer writing is not close yet. Then continue only when buffer
writing completed and close the channel to signal completion. This
can make sure all the writer buffer are send out in sequence as
what their order during pending buffering and also make better use
of CPU cycle usage. The sequence of writer can only be sure if the
buffer writer executed in the same goroutine.

Update the test for reflect this commit change which tests for
sequence pending writer.